Implement notifications
This makes it possible to receive notifications (via email) when a patch
has been updated.

It's possible to subscribe to notifications on a specific patch. It is also
possible to set the user profile up so that one receives notifications for
all patches where one is author, reviewer or committer on.

Notifications are, of course, only sent when other people make modifications
to the entry.