Actually deliver CCed emails to patch authors
Previously we would set the Cc header of the email, but not actually
send a copy to the user. That way if the user was not subscribed to the
hackers email, they'd end up not receiving the email at all, which
somewhat defeats one of the original points of adding the Cc in the
first place.

Spotted by Peter Geoghegan