Support local cfbot integration development with dummy data (#8)
This adds some basic functionality to allow making changes to the CFBot
integration locally, without having to push to prod/staging. To do so
this adds the CFBot integration tables to the admin panel and adds some
dummy CFBot data to our dummy development dataset. One somewhat
interesting thing this required that now the `CfbotBranch` has an
overridden `save` method, which stores `NULL` in the commit_id if an
empty string is saved. This only impacts the save behaviour of the
Django admin interface because the webhook that the actual CFBot calls
bypasses the Django models and inserts directly into the underlying
tables.