Implement cache busting for our own static files (#23)
Browsers cache static files. That's obviously good. Currently it results
in our we change files server side, users don't always see them.

This generates a GET parameter that's unique per startup of the python
process to bust the cache of the client, so that it pulls in possibly
updated JS/CSS files. Obviously this is a bit more often than necessary,
but caching should still work most of the time and our custom css and js
files are tiny anyway. At least the changes that are made will now be
reflected to our users.