Default author to self when adding a new patch (#26)
People sometimes forget to configure the author field when adding a
patch to the commitfest. It's also kinda silly to always have to search
for your own name. In the cases that people add commitfest entries for
patches of others, removing themselves and adding the other person is
only a backspace presses away.

Apart from that this also updates the patch creation form to include all
the fields from the patch edit form. Instead of having the patch
creation submit button redirect to the patch edit page.

Fixes #14