Autofill new patch description based on selected thread (#30)
It's a bit silly that you have to manually enter something in the
description, when you just selected a thing with that title from the
thread search modal. So this starts auto-filling the description. This
only happens when the field is empty, and will strip some common
prefixes like `Re:` and `[PATCH]`.

Fixes #16