Fix reverse sorting
The unsafe fixes made by the biome linter actually changed behaviour in
an unintended way. It had replaced `==` with `===`, which has much more
sensible behaviour. But because the value of the input was of type
String and we were comparing that to a Number the comparison now always
failed. This fixes this regression by changing the code to explicitly
parse the string as a number, instead of using the `==` comparison.

The reason I hadn't noticed after applying the linter fixes is because
my cache was not busted by the Django dev server, because it did not
restart on JS only changes. This will be fixed in a follow up commit.