Reload dev uwsgi on JS and CSS changes
Reloading our dev uwsgi isn't necessary to serve the correct file JS and
CSS file. But the cache busting employed in 8d3802aec56 only happens
when the server restarts. So by not reloading browsers would still cache
the old version.

This problem was found  because a bug got in when I made some final
modifications to 796c427b3cf. That bug was fixed in 5dfc380105f.