WIP: add a Drafts commitfest
Drafts is a special always-open commitfest for the purpose of holding
and testing draft patch submissions. Having such a holding area should
make it easier for people to keep track of patchsets they're not quite
ready to submit for review.

Internally, Drafts is assigned a static ID of zero. This is chosen
because a) it does not overlap with Django's default AutoField sequence,
which begins at one, and b) it requires no updates to the current URL
patterns, which match nonnegative integers.

The Drafts entry has the special status STATUS_DRAFT so that it does not
conflict with pre-existing coded assumptions on what "open", "future",
etc. mean. STATUS_DRAFT CFs are excluded from the "num_cfs" count for a
patch.

The new /close/draft handler is added to swap patches into Drafts.
Patches are then removed by moving them to the next open CF, or by
closing as usual.

Prior to this patch:
- CFs with IDs less than the current in-progress CF could safely be
  assumed closed,
- patches only ever moved forward through increasing CF IDs, and
- the latest CF start date determined a patch's "current" CF.

These assumptions all break under the current model. They have been
modified:
- use STATUS_CLOSED specifically when deciding whether a CF is closed
- when moving a patch between CFs, allow for the possibility of an
  existing entry in the junction table
- a patch's "current" CF is determined by its latest entry date

TODO:
- ensure all prior assumptions on CF ID are cleaned up