Allow sorting commitfest entries by "failing since" (#47)
This adds a new column to the `CfbotBranch` model: `failing_since`. This
tracks when a patch first started to fail the CI or started requiring a
rebase. This value its primary use is to allow people to sort by the
"CI status" column in the patch list. This allows people to quickly see
what patches are healthy, which require the author to take action, and
which might benefit from a reviewer pinging the author that the patch
has been broken for a while. Non-failing builds are sorted on their CI
build creation time, so patches that are currently running CI are shown
at the top, followed by patches for which CI succeeded most recently.

The values of needs_rebase_since and failing_since are also shown on
hover and on the patch pages.

Fixes #42