Remove latest activity column
Having both "Latest email" and "Latest activity" on the main commitfest
list kind of useless. The "Latest activity" often is simply the date
that a patch was moved to the next commitfest, so it's content is rather
useless in general. If we improve the commitfest app to contain more
useful info (e.g. by implementing #19) it might become more useful to
have this timestamp though. For now it seems fine to remove it though,
and once we implement #10 we can add his back as an option.