Remove latest activity column (#49)
Having both "Latest email" and "Latest activity" as a column on the
commitfest page is kind of useless. The "Latest activity" often is
simply the date that a patch was moved to the next commitfest, so it's
content is rather useless in general. If we improve the commitfest app
to contain more useful info, then it might become more useful to have
this timestamp. For now it seems fine to remove it though. We can always
add it back later if it turns out that people miss it.

Fixes #21