Show "Latest email" column as a relative time (#50)
Exact dates of the last email are usually not super useful when looking
at the commitfest page. This instead starts showing a relative time
since. For the rare cases where the exact date & time is needed, the
tooltip will show the exact date & time. When this gets deployed to prod
a small change also needs to be made to the cfbot, because this breaks
the cfbot its html scraping.