Revamp the homepage to show you patches you might be interested in (#51)
The homepage was not showing a lot of useful information. It had a list
of mostly ancient commitfests, a search box, and a few links to more
useful pages. This revamps the homepage with an initial version of a
personalized dashboard. The dashboard shows patches that the user is
subscribed to, either because they are marked as an author or a
reviewer. We then try to group and sort these patches in an attempt to
let the user easily find the patches that require their attention.

Many things can likely be tweaked/improved. I think this first version
is already a huge improvement over the status quo, so I think it's worth
committing deploying this mostly as is and then iterating on it based on
feedback.

This is what it looks like:
![screenshot of
homepage](https://github.com/user-attachments/assets/6f464597-30f9-4f4b-9b98-17ac90b64b60)

Fixes #36