Use selectize for author/reviewer filter for patch lists
This has a few benefits:
1. It makes it much easier to select a user, since you don't have to
   scroll through the very long list anymore.
2. It makes the pageload faster, because the full user list is not part
   of the generated html.
3. No heavy query to get all the users.

Based on some previous quick attempts, this is expected to make the
page load significantly faster.

By doing this I also realized that the Author multiselect for patch
create/edit was showing "username (firstname lastname)" instead of
"firstname lastname (username)" like we do everywhere else. So this
fixes that too.