Patch UX Enhancement Refactor
In the interest of breaking up this major change to both UX, backend,
and Workflow this patch separates out the main UX redesign choices for
independent consideration.

There is a new workflow display that will be polished and described
once the final components have been incorporated.  For now it simply
places all of the action menu items on the screen; and incorporates
a more coordinated and flow-oriented layout for the information
in the existing key-value table.

The #history-table is hidden by default but does include potentially
useful information and so is expected to remain long-term.  The status
history field in the keyvalue-table may be moved here at some point.

The elements from the #keyvalue-table that did not get incorporated
into the workflow display are placed into a new workflow supplemental
table.

Unlike the key-value table, the new tables are horizontally compact.

The #keyvalue-table is presently hidden but viewable.  Conceptually,
the contents of this table should never be needed in this view and
at some point the table itself can be removed.  For now it is left
as a fallback.

The existing action bar has been renamed the administrative bar.
This will be visible only to staff and has minimal logic for showing
and hiding UI widgets.  That way, should the UI logic be faulty, it
is generally still possible for staff to fix it.  In this commit the
contents of the bar remain unchanged, only the file name is different.