Migration 'next' close action to transitions
Remove all usage of close/next in the UI (just administrative actions).

Remove all 'next' detection logic.  The UI will provide valid options
to the user to choose between and they can make an informed choice;
especially since for the typical user there will only be a single choice,
to the Future Commitfest or back to the Parked one.  Committers can
also choose to bring any patch into the In Progress or Open/Interlude.

The implicit constraint 'next' applied to the system was that,
practically speaking, as Patches moved they always move to a Commitfest
that was started in the future: thus current_patch_on_commitfest could
just use a descending start date sort to find the primary Commitfest.
The transition action uses the Workflow implementation which allows
for a Patch to re-enter the same Commitfest.  This changes the
entrydate on the POC which is now used to find the current Commitfest.

The Workflow module tries to improve upon the History messages being
generated.  Additional thoughts forthcoming in this area.  Should be
simple enough to retain the status quo on these if desired, especially
since I'm disliking the absence of a cfid column; which is why i added
the relevant POC Commitfest name to the messages.

The new code is a bit chatty in terms of feedback to the user.  Though,
on the flip side, it intentionally doesn't feel the need to verify
that the button just clicked was intended.  That needs to be tidied
up based upon the guidelines I need to review.