Protect actions behind user.is_authenticated
The idea of having actions visible and then redirect for authentication
is acceptable in general but the /patch view is now driven by the
role of the user viewing the page that it seems easier to simply show
no actions if the user is not authenticated and add a note just above
the action table indicating the read-only mode.

Add missing is_staff protection for the administrative actions form.

Just outright removing the key-value table if you are not logged in.
Don't want to edit it, want to get rid of it.