Introduce Workflow, Drafts, and API
The Commitfest Workflow is a formalization and extension of the existing
commitfest-only protocol.  There is now some in-app documentation and the
idea of moving patches to the "next CF" has begun being deprecated. Instead
the UI provides the user valid choices on where to transition patches,
which, with the introduction of the Drafts commitfest, is basically one
or the other.  Committers gain the ability to transition patches into
In Progress if desired.  Patches within In Progress commitfests can
be moved into either Open or Drafts.  Begin enforcement of a maximum
of one commifest in each of  Parked, Open, and In Progress.

The workflow defines the drafts commitfest as lasting for one year
matching up with the major release process.  Namely, once the
commifest leading up to feature freeze is In Progress a new
Drafts commitfest should be created.

With the workflow changes it is necessary for CFBot to be made aware
of the specific commitfest ids that are active.  Provide a JSON API
for this purpose to begin weaning CFBot off scraping html.

The workflow in-app documentation includes a month-based schedule
showing which commifests are active, and their names, in each month
(among other related details).  Begin enforcing the existing rule
that future commitfests cannot have patches via triggers.  Future
work on the administrative and UX aspects of the CFApp likely will
do away with future commitfests altogether.  Thoughts on the topic
are welcome in the GitHub issues.