Make committer select box easier to use
The select box for selecting a committer of a patch was kinda hard to
use. This fixes it in two ways:
1. Sorts names by first names, instead of last names. This way the list
   actually looks ordered.
2. Use Selectize to make the select box searchable.

To make such a change to selectize simpler in the future this starts
including the selectize css/js on every page and adds a simple
"enable-selectize" class that can be used to easily enable selectize on
any select box.