Add Tags for patches
A Tag is an arbitrary label for a patch in the Commitfest UI. Other than
helping users identify patches of interest, it has no other semantic
meaning to the CF application.

Tags are created using the administrator interface. They consist of a
unique name and a background color. The color should be sent from
compliant browsers in #rrggbb format, which is stored without backend
validation; to avoid CSS injection, any non-conforming values are
replaced with black during templating.