Make user dropdowns searchable when not logged in
The endpoint required login for afaict no particularly good reason. All
it returns are usernames and names of users in the system. These are
already available when browsing on many pages. Also a login requirement
for this doesn't seem a particularly high hurdle to overcome as it is
easy to make an account.